pyproject.toml
[project]
name = "mi-alcohol-persona"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.38.0",
    "beautifulsoup4>=4.12.3",
    "distilabel>=1.4.1",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
]
